Refraction needs double to solve last artifacts :-( Hopefully this is the final fix...

+++ src/core/RefractionExtinction.cpp
@@ -114,14 +114,14 @@ void Refraction::innerRefractionForward(Vec3d& altAzPos) const
 	const double length = altAzPos.length();
 	const double sinGeo = altAzPos[2]/length;
 	double geom_alt_rad = std::asin(sinGeo);
-	float geom_alt_deg = 180./M_PI*geom_alt_rad;
+	double geom_alt_deg = 180./M_PI*geom_alt_rad;
 	if (geom_alt_deg > MIN_GEO_ALTITUDE_DEG)
 	{
 		// refraction from Saemundsson, S&T1986 p70 / in Meeus, Astr.Alg.
-		float r=press_temp_corr * ( 1.02f / std::tan((geom_alt_deg+10.3f/(geom_alt_deg+5.11f))*M_PI/180.f) + 0.0019279f);
+		double r=press_temp_corr * ( 1.02 / std::tan((geom_alt_deg+10.3/(geom_alt_deg+5.11))*M_PI/180.) + 0.0019279);
 		geom_alt_deg += r;
-		if (geom_alt_deg > 90.f)
-			geom_alt_deg=90.f;
+		if (geom_alt_deg > 90.)
+			geom_alt_deg=90.;
 	}
 	else if(geom_alt_deg>MIN_GEO_ALTITUDE_DEG-TRANSITION_WIDTH_GEO_DEG)
 	{
@@ -136,7 +136,6 @@ void Refraction::innerRefractionForward(Vec3d& altAzPos) const
 	const double refr_alt_rad=geom_alt_deg*M_PI/180.;
 	const double sinRef=std::sin(refr_alt_rad);
 	const double shortenxy=std::sqrt((1.-sinRef*sinRef)/(1.-sinGeo*sinGeo)); // we need double's mantissa length here, sorry!
-	//if (sinGeo>0) Q_ASSERT(shortenxy<=1.0); // usual case: should have pushed vector upward.
 
 	altAzPos[0]*=shortenxy;
 	altAzPos[1]*=shortenxy;
@@ -149,11 +148,11 @@ void Refraction::innerRefractionBackward(Vec3d& altAzPos) const
 	// going from observed position/magnitude to geometrical position and atmosphere-free mag.
 	const double length = altAzPos.length();
 	const double sinObs = altAzPos[2]/length;
-	float obs_alt_deg=180./M_PI*std::asin(sinObs);
+	double obs_alt_deg=180./M_PI*std::asin(sinObs);
 	if (obs_alt_deg > 0.22879f)
 	{
 		// refraction from Bennett, in Meeus, Astr.Alg.
-		float r=press_temp_corr * (1.f / std::tan((obs_alt_deg+7.31f/(obs_alt_deg+4.4f))*M_PI/180.f) + 0.0013515f);
+		double r=press_temp_corr * (1. / std::tan((obs_alt_deg+7.31/(obs_alt_deg+4.4))*M_PI/180.) + 0.0013515);
 		obs_alt_deg -= r;
 	}
 	else if (obs_alt_deg > MIN_APP_ALTITUDE_DEG)
@@ -178,7 +177,6 @@ void Refraction::innerRefractionBackward(Vec3d& altAzPos) const
 	const double geo_alt_rad=obs_alt_deg*M_PI/180.;
 	const double sinGeo=std::sin(geo_alt_rad);
 	const double longerxy=std::sqrt((1.-sinGeo*sinGeo)/(1.-sinObs*sinObs));
-	//if (sinGeo>0) Q_ASSERT(longerxy>=1.); // // usual case: should have pushed vector downward.
 	altAzPos[0]*=longerxy;
 	altAzPos[1]*=longerxy;
 	altAzPos[2]=sinGeo*length;
